+++ internal/containerrunner/init_test.go
@@ -1,0 +1,13 @@
+package containerrunner_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestContainerRunner(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Container Runner")
+}