+++ pkg/types/imagerunnerinput.go
@@ -3,6 +3,8 @@ package types
 import (
 	"fmt"
 	"regexp"
+
+	"github.com/docker/docker/api/types/container"
 )
 
 var cleaningRegex *regexp.Regexp
@@ -36,4 +38,26 @@ func (i *ImageRunnerInput) Name() string {
 		i.name = fmt.Sprintf("%s-%s", cleanText(i.Topic), cleanText(i.FunctionMetadata.Image))
 	}
 	return i.name
+}
+
+//ContainerConfig ...
+func (i *ImageRunnerInput) ContainerConfig(defaultEnv []string) *container.Config {
+	env := append(defaultEnv, fmt.Sprintf("FIRED_BY=%s", i.Topic))
+	return &container.Config{
+		AttachStdin:  true,
+		AttachStdout: true,
+		AttachStderr: true,
+		Image:        i.FunctionMetadata.Image,
+		OpenStdin:    true,
+		Tty:          false,
+		StdinOnce:    true,
+		Labels: map[string]string{
+			"mqttfaas_runtime": i.Name(),
+			"mqtt_topic":       i.Topic,
+		},
+		Volumes: map[string]struct{}{
+			"/data": *new(struct{}),
+		},
+		Env: env,
+	}
 }